rename(auth): auth 폴더 구조
auth 패키지를 domain 패키지로 구조를 변경했습니다.
