+++ src/main/java/com/space/server/common/jwt/filter/LoginFilter.java
@@ -1,7 +1,7 @@
 package com.space.server.common.jwt.filter;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.space.server.common.auth.service.dto.CustomUserDetails;
+import com.space.server.domain.auth.service.dto.CustomUserDetails;
 import com.space.server.common.jwt.dto.LoginRequest;
 import com.space.server.common.jwt.util.JwtUtil;
 import com.space.server.domain.user.domain.value.Role;
