+++ src/main/java/com/space/server/common/jwt/util/JwtUtil.java
@@ -1,7 +1,7 @@
 package com.space.server.common.jwt.util;
 
-import com.space.server.common.auth.domain.Refresh;
-import com.space.server.common.auth.domain.repository.RefreshRepository;
+import com.space.server.domain.auth.domain.Refresh;
+import com.space.server.domain.auth.domain.repository.RefreshRepository;
 import com.space.server.common.jwt.exception.ExpiredRefreshTokenException;
 import com.space.server.common.jwt.exception.ExpiredTokenException;
 import com.space.server.domain.user.domain.value.Role;
